.history/requirements_20260401230258.txt
google-genai
flask
pandas
requests
gunicorn